+++ mlxtend/_base/_cluster.py
@@ -1,4 +1,4 @@
-# Sebastian Raschka 2014-2018
+# Sebastian Raschka 2014-2019
 # mlxtend Machine Learning Library Extensions
 #
 # Base Clusteer (Clutering Parent Class)