+++ mlxtend/plotting/plot_linear_regression.py
@@ -1,4 +1,4 @@
-# Sebastian Raschka 2014-2018
+# Sebastian Raschka 2014-2019
 # mlxtend Machine Learning Library Extensions
 #
 # Function for plotting linear regression fits via scikit-learn and matplotlib.