+++ mlxtend/preprocessing/copy_transformer.py
@@ -1,4 +1,4 @@
-# Sebastian Raschka 2014-2018
+# Sebastian Raschka 2014-2019
 # mlxtend Machine Learning Library Extensions
 #
 # A Class that returns a copy of a dataset in a scikit-learn pipeline.