+++ docs/sources/CHANGELOG.md
@@ -16,7 +16,7 @@ The CHANGELOG for the current development version is available at
 
 ##### New Features
 
-- -
+- Adds a new function, `mlxtend.evaluate.bias_variance_decomp` that decomposes the loss of a regressor or classifier into bias and variance terms ([#470](https://github.com/rasbt/mlxtend/pull/470))
 
 ##### Changes