+++ docs/sources/CHANGELOG.md
@@ -23,10 +23,10 @@ The CHANGELOG for the current development version is available at
 
 - Changed the default solver in `PrincipalComponentAnalysis` to `'svd'` instead of `'eigen'` to improve numerical stability. ([#474](https://github.com/rasbt/mlxtend/pull/474))
 
+
 ##### Bug Fixes
 
-- The eigenvectors maybe have not been sorted in certain edge cases if solver was `'eigen'` in `PrincipalComponentAnalysis`. ([#477](https://github.com/rasbt/mlxtend/pull/477))
-
+- The eigenvectors maybe have not been sorted in certain edge cases if solver was `'eigen'` in `PrincipalComponentAnalysis` and `LinearDiscriminantAnalysis`. ([#477](https://github.com/rasbt/mlxtend/pull/477), [#478](https://github.com/rasbt/mlxtend/pull/478))
 
 
 ### Version 0.14.0 (11-09-2018)